refactor: mail-session creation to own unit
